+++ HISTORY.md
@@ -16,6 +16,7 @@
 * locale: cleaned up some ambiguous syntax
 * gui: added tutorial scene launch and escape methods
 * German translation update for new "launch..." string (for tutorial scenes)
+* gui: update visibility of scene icons when launching or escaping a tutorial
 
 ### 2025-10-29
 * gui: made tutorial dialog resizable